inject csv manager as dependency
